Refactor api tests
