+++ package.json
@@ -38,7 +38,9 @@
         "codacy-coverage": "^3.4.0",
         "coveralls": "^3.0.2",
         "mocha": "^5.2.0",
+        "mocha-prepare": "^0.1.0",
         "mockgoose": "^8.0.1",
+        "mongo-unit": "^1.4.4",
         "nyc": "^13.2.0"
     }
 }
