+++ test/api/suites/member_test.js
@@ -1,34 +1,13 @@
-process.env.NODE_ENV = 'test';
-
-let mongoose = require('mongoose');
-
-let Mockgoose = require('mockgoose').Mockgoose;
-let mockgoose = new Mockgoose(mongoose);
-
-let chai = require('chai');
-let chaiHttp = require('chai-http');
-let server = require('../server');
-let should = chai.should();
+const chai = require('chai');
+const chaiHttp = require('chai-http');
+const server = require('../../../server');
+const should = chai.should();
+const expect = chai.expect;
+const assert = chai.assert;
 
 chai.use(chaiHttp);
 
-before(function(done) {
-    mockgoose.prepareStorage().then(function() {
-        mongoose.connect(
-            'mongodb://example.com/TestingDB',
-            { useNewUrlParser: true },
-            function(err) {
-                done(err);
-            }
-        );
-    });
-});
-
-after(function() {
-    process.exit(0);
-});
-
-describe('Members suite /ADD,/GET,/GET/:ID,/EDIT/:ID, /EDIT,/DELETE', () => {
+module.exports = function() {
     it('it should add a new member', done => {
         let testMember1 = {
             firstName: 'Charles',
@@ -203,4 +182,4 @@ describe('Members suite /ADD,/GET,/GET/:ID,/EDIT/:ID, /EDIT,/DELETE', () => {
                 done();
             });
     });
-});
+};
