+++ test/mocha.opts
@@ -1,1 +1,5 @@
 --timeout 120000
+--full-trace
+--recursive
+--require test/api/prepareMongoUnit
+--exit
